refactor(24140): refactor nesting of component for testing
