+++ hivemq-edge/src/frontend/src/extensions/datahub/components/nodes/NodeWrapper.tsx
@@ -1,6 +1,6 @@
 import { FC, ReactNode, useEffect, useMemo, useState } from 'react'
 import { useLocation, useNavigate } from 'react-router-dom'
-import { NodeProps } from 'reactflow'
+import { NodeProps, NodeToolbar, Position } from 'reactflow'
 import { Avatar, BoxProps, Card, CardBody, CardBodyProps, HStack } from '@chakra-ui/react'
 
 import NodeDatahubToolbar from '@datahub/components/nodes/NodeDatahubToolbar.tsx'
@@ -8,6 +8,7 @@ import { DataHubNodeData, PolicyDryRunStatus } from '@datahub/types.ts'
 import { getDryRunStatusIcon } from '@datahub/utils/node.utils.ts'
 import { parseHotkey } from '@datahub/utils/hotkeys.utils.ts'
 import { DATAHUB_HOTKEY } from '@datahub/utils/datahub.utils.ts'
+import useDataHubDraftStore from '@datahub/hooks/useDataHubDraftStore.ts'
 
 interface NodeWrapperProps extends NodeProps<DataHubNodeData> {
   children: ReactNode
@@ -16,6 +17,7 @@ interface NodeWrapperProps extends NodeProps<DataHubNodeData> {
 }
 
 export const NodeWrapper: FC<NodeWrapperProps> = ({ selected, children, route, wrapperProps, data }) => {
+  const { nodes } = useDataHubDraftStore()
   const [internalSelection, setInternalSelection] = useState(false)
   const navigate = useNavigate()
   const { pathname } = useLocation()
@@ -25,6 +27,11 @@ export const NodeWrapper: FC<NodeWrapperProps> = ({ selected, children, route, w
   useEffect(() => {
     if (!selected) setInternalSelection(false)
   }, [selected])
+
+  const isSingleSelect = useMemo(() => {
+    const selectedNodes = nodes.filter((node) => node.selected)
+    return selectedNodes.length === 1
+  }, [nodes])
 
   const onHandleCopy = () => {
     document.dispatchEvent(new KeyboardEvent('keydown', parseHotkey(DATAHUB_HOTKEY.COPY)))
@@ -61,12 +68,9 @@ export const NodeWrapper: FC<NodeWrapperProps> = ({ selected, children, route, w
   return (
     <>
       {isToolbarEnabled && (
-        <NodeDatahubToolbar
-          isVisible={selected}
-          onCopy={onHandleCopy}
-          onDelete={onHandleDelete}
-          onEdit={onHandleEdit}
-        />
+        <NodeToolbar isVisible={selected && isSingleSelect} position={Position.Top} offset={8}>
+          <NodeDatahubToolbar onCopy={onHandleCopy} onDelete={onHandleDelete} onEdit={onHandleEdit} />
+        </NodeToolbar>
       )}
       <Card
         variant="elevated"
